requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
coverage==5.5
dj-database-url==0.5.0
Django==3.2.5
django-coverage-plugin==2.0.0
django-crispy-forms==1.12.0
django-currentuser==0.5.3
django-embed-video==1.4.0
django-heroku==0.3.1
django-shapeshifter==18.9.23
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dateutil==1.5
python-decouple==3.4
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
